Split server.rs into server.rs, standby.rs and member.rs
This commit splits the very large server.rs file into standby.rs and member.rs
which live in the server module.
